parsey
Trait Parser

pub trait Parser<Token, Error>: Iterator<Item = Token> + Sized {
    type Root: Ast<Token, Error>;

    // Required method
    fn expect(
        peekable_parser: &mut PeekableParser<Self, Token, Error>,
        expected: Token,
    ) -> Result<(), Error>;

    // Provided method
    fn parse(self) -> Result<Self::Root, Error> { ... }
}

A trait representing a generic parser that consumes tokens and produces an AST.

This trait provides an abstraction for parsers that process tokens.

Type Parameters

Token: The type of tokens being parsed.

Error: The type of errors that may occur during parsing.

Required Associated Types

type Root: Ast<Token, Error>

The root type of the AST produced by this parser.

Required Methods

fn expect( peekable_parser: &mut PeekableParser<Self, Token, Error>, expected: Token, ) -> Result<(), Error>

Validates whether a given token matches the expected token.

This method is used to verify that the next token in the parsing sequence matches what is expected according to the grammar rules. if it matches, the token is consumed.

Parameters

token: The token to validate against the expected token

Returns

Ok(()) if the token matches the expected token

Err(Error) if the token does not match the expected token

Provided Methods

fn parse(self) -> Result<Self::Root, Error>

Parses an AST from a peekable token stream.

Parameters

parser: A peekable iterator implementing the Parser trait.

Returns

Returns the parsed AST root node or an error if parsing fails.

Errors

Returns an error of type Error if the token sequence does not match the expected structure.

Object Safety

This trait is not object safe.